:truck: Rename/move AsyncTestCase into core pkg
