+++ src/tests/fixtures/test_load_fixtures.py
@@ -1,6 +1,6 @@
 from apps.hero.models import Hero
 from core.db.fixtures import LoadFixtures
-from tests.base import AsyncTestCase
+from core.test.async_case import AsyncTestCase
 
 
 class TestLoadFixture(AsyncTestCase):
