+++ src/tests/services/test_db_queries.py
@@ -1,5 +1,5 @@
 from apps.hero.models import Hero
-from tests.base import AsyncTestCase
+from core.test.async_case import AsyncTestCase
 
 BASE_URL = "http://test"
